LayoutTests/imported/w3c/web-platform-tests/cookie-store/idlharness.tentative.https.any-expected.txt


PASS idl_test setup
PASS idl_test validation
PASS Partial interface ServiceWorkerRegistration: original interface defined
PASS Partial interface ServiceWorkerRegistration: valid exposure set
PASS Partial interface ServiceWorkerRegistration: member names are unique
PASS Partial interface Window: original interface defined
PASS Partial interface Window: member names are unique
PASS Partial interface ServiceWorkerGlobalScope: original interface defined
PASS Partial interface ServiceWorkerGlobalScope: member names are unique
PASS Partial interface Window[2]: member names are unique
PASS Window includes GlobalEventHandlers: member names are unique
PASS Window includes WindowEventHandlers: member names are unique
PASS Window includes WindowOrWorkerGlobalScope: member names are unique
PASS WorkerGlobalScope includes WindowOrWorkerGlobalScope: member names are unique
PASS Window includes AnimationFrameProvider: member names are unique
PASS Window includes WindowSessionStorage: member names are unique
PASS Window includes WindowLocalStorage: member names are unique
FAIL CookieStore interface: existence and properties of interface object assert_own_property: self does not have own property "CookieStore" expected property "CookieStore" missing
FAIL CookieStore interface object length assert_own_property: self does not have own property "CookieStore" expected property "CookieStore" missing
FAIL CookieStore interface object name assert_own_property: self does not have own property "CookieStore" expected property "CookieStore" missing
FAIL CookieStore interface: existence and properties of interface prototype object assert_own_property: self does not have own property "CookieStore" expected property "CookieStore" missing
FAIL CookieStore interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "CookieStore" expected property "CookieStore" missing
FAIL CookieStore interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "CookieStore" expected property "CookieStore" missing
FAIL CookieStore interface: operation get(USVString) assert_own_property: self does not have own property "CookieStore" expected property "CookieStore" missing
FAIL CookieStore interface: operation get(optional CookieStoreGetOptions) assert_own_property: self does not have own property "CookieStore" expected property "CookieStore" missing
FAIL CookieStore interface: operation getAll(USVString) assert_own_property: self does not have own property "CookieStore" expected property "CookieStore" missing
FAIL CookieStore interface: operation getAll(optional CookieStoreGetOptions) assert_own_property: self does not have own property "CookieStore" expected property "CookieStore" missing
FAIL CookieStore interface: operation set(USVString, USVString) assert_own_property: self does not have own property "CookieStore" expected property "CookieStore" missing
FAIL CookieStore interface: operation set(CookieInit) assert_own_property: self does not have own property "CookieStore" expected property "CookieStore" missing
FAIL CookieStore interface: operation delete(USVString) assert_own_property: self does not have own property "CookieStore" expected property "CookieStore" missing
FAIL CookieStore interface: operation delete(CookieStoreDeleteOptions) assert_own_property: self does not have own property "CookieStore" expected property "CookieStore" missing
FAIL CookieStore interface: attribute onchange assert_own_property: self does not have own property "CookieStore" expected property "CookieStore" missing
FAIL CookieStore must be primary interface of self.cookieStore assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL Stringification of self.cookieStore assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CookieStore interface: self.cookieStore must inherit property "get(USVString)" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CookieStore interface: calling get(USVString) on self.cookieStore with too few arguments must throw TypeError assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CookieStore interface: self.cookieStore must inherit property "get(optional CookieStoreGetOptions)" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CookieStore interface: calling get(optional CookieStoreGetOptions) on self.cookieStore with too few arguments must throw TypeError assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CookieStore interface: self.cookieStore must inherit property "getAll(USVString)" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CookieStore interface: calling getAll(USVString) on self.cookieStore with too few arguments must throw TypeError assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CookieStore interface: self.cookieStore must inherit property "getAll(optional CookieStoreGetOptions)" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CookieStore interface: calling getAll(optional CookieStoreGetOptions) on self.cookieStore with too few arguments must throw TypeError assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CookieStore interface: self.cookieStore must inherit property "set(USVString, USVString)" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CookieStore interface: calling set(USVString, USVString) on self.cookieStore with too few arguments must throw TypeError assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CookieStore interface: self.cookieStore must inherit property "set(CookieInit)" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CookieStore interface: calling set(CookieInit) on self.cookieStore with too few arguments must throw TypeError assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CookieStore interface: self.cookieStore must inherit property "delete(USVString)" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CookieStore interface: calling delete(USVString) on self.cookieStore with too few arguments must throw TypeError assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CookieStore interface: self.cookieStore must inherit property "delete(CookieStoreDeleteOptions)" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CookieStore interface: calling delete(CookieStoreDeleteOptions) on self.cookieStore with too few arguments must throw TypeError assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CookieStore interface: self.cookieStore must inherit property "onchange" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CookieStoreManager interface: existence and properties of interface object assert_own_property: self does not have own property "CookieStoreManager" expected property "CookieStoreManager" missing
FAIL CookieStoreManager interface object length assert_own_property: self does not have own property "CookieStoreManager" expected property "CookieStoreManager" missing
FAIL CookieStoreManager interface object name assert_own_property: self does not have own property "CookieStoreManager" expected property "CookieStoreManager" missing
FAIL CookieStoreManager interface: existence and properties of interface prototype object assert_own_property: self does not have own property "CookieStoreManager" expected property "CookieStoreManager" missing
FAIL CookieStoreManager interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "CookieStoreManager" expected property "CookieStoreManager" missing
FAIL CookieStoreManager interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "CookieStoreManager" expected property "CookieStoreManager" missing
FAIL CookieStoreManager interface: operation subscribe(sequence<CookieStoreGetOptions>) assert_own_property: self does not have own property "CookieStoreManager" expected property "CookieStoreManager" missing
FAIL CookieStoreManager interface: operation getSubscriptions() assert_own_property: self does not have own property "CookieStoreManager" expected property "CookieStoreManager" missing
FAIL CookieStoreManager interface: operation unsubscribe(sequence<CookieStoreGetOptions>) assert_own_property: self does not have own property "CookieStoreManager" expected property "CookieStoreManager" missing
FAIL CookieStoreManager must be primary interface of self.registration.cookies assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL Stringification of self.registration.cookies assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CookieStoreManager interface: self.registration.cookies must inherit property "subscribe(sequence<CookieStoreGetOptions>)" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CookieStoreManager interface: calling subscribe(sequence<CookieStoreGetOptions>) on self.registration.cookies with too few arguments must throw TypeError assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CookieStoreManager interface: self.registration.cookies must inherit property "getSubscriptions()" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CookieStoreManager interface: self.registration.cookies must inherit property "unsubscribe(sequence<CookieStoreGetOptions>)" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CookieStoreManager interface: calling unsubscribe(sequence<CookieStoreGetOptions>) on self.registration.cookies with too few arguments must throw TypeError assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CookieChangeEvent interface: existence and properties of interface object assert_own_property: self does not have own property "CookieChangeEvent" expected property "CookieChangeEvent" missing
FAIL CookieChangeEvent interface object length assert_own_property: self does not have own property "CookieChangeEvent" expected property "CookieChangeEvent" missing
FAIL CookieChangeEvent interface object name assert_own_property: self does not have own property "CookieChangeEvent" expected property "CookieChangeEvent" missing
FAIL CookieChangeEvent interface: existence and properties of interface prototype object assert_own_property: self does not have own property "CookieChangeEvent" expected property "CookieChangeEvent" missing
FAIL CookieChangeEvent interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "CookieChangeEvent" expected property "CookieChangeEvent" missing
FAIL CookieChangeEvent interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "CookieChangeEvent" expected property "CookieChangeEvent" missing
FAIL CookieChangeEvent interface: attribute changed assert_own_property: self does not have own property "CookieChangeEvent" expected property "CookieChangeEvent" missing
FAIL CookieChangeEvent interface: attribute deleted assert_own_property: self does not have own property "CookieChangeEvent" expected property "CookieChangeEvent" missing
FAIL CookieChangeEvent must be primary interface of new CookieChangeEvent("change") assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: CookieChangeEvent"
FAIL Stringification of new CookieChangeEvent("change") assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: CookieChangeEvent"
FAIL CookieChangeEvent interface: new CookieChangeEvent("change") must inherit property "changed" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: CookieChangeEvent"
FAIL CookieChangeEvent interface: new CookieChangeEvent("change") must inherit property "deleted" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: CookieChangeEvent"
PASS ExtendableCookieChangeEvent interface: existence and properties of interface object
FAIL ServiceWorkerRegistration interface: attribute cookies assert_true: The prototype object must have a property "cookies" expected true got false
FAIL ServiceWorkerRegistration interface: self.registration must inherit property "cookies" with the proper type assert_inherits: property "cookies" not found in prototype chain
FAIL Window interface: attribute cookieStore assert_own_property: The global object must have a property "cookieStore" expected property "cookieStore" missing
FAIL Window interface: self must inherit property "cookieStore" with the proper type assert_own_property: expected property "cookieStore" missing
